refactor: Move various app settings into settings module
